String.prototype.localeCompare is worthless.
The results of this function are implementation dependent so, fuck using
it at all. While there is a new implementation accepting a couple more
arguments to specify the locale, I haven't researched it enough to know
whether or not these new extra arguments are standardized and
implemented anywhere besides google chrome. If I get verification that
there is some standard so I can expect the same output on all platforms
then I'll implement this function again. In the meantime it's going to
throw an error because I'm too lazy to add in static arrays of locale
charsets and implement my own version of this function. I had no idea
that "alphabetical order" was subject to interpretation. Seems like
whatever dictionary is considered the standard for a locale would have
already standardized the sort order for words in that locale and we
shouldn't have this issue in JavaScript at all...